telegram_bot/requirements.txt
python-telegram-bot==20.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.3
python-dotenv==1.0.0